Move extractors into individual modules
