+++ tracecat/expressions/ioc_extractors/email.py
@@ -1,0 +1,55 @@
+import re
+
+from pydantic import BaseModel, EmailStr, ValidationError
+
+# EMAIL
+EMAIL_REGEX = r"[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}"
+
+
+class EmailModel(BaseModel):
+    email: EmailStr
+
+
+def normalize_email(email: str) -> str:
+    """Convert sub-addressed email to a normalized email address.
+
+    This function:
+    1. Converts the email to lowercase
+    2. Removes the subaddress part (everything after + in the local part)
+
+    Example: User.Name+Newsletter@Example.COM -> user.name@example.com
+    """
+    email = email.lower()
+    local_part, domain = email.split("@")
+    local_part = local_part.split("+")[0]
+    return f"{local_part}@{domain}"
+
+
+def extract_emails(text: str, normalize: bool = False) -> list[str]:
+    """Extract unique emails from a string.
+
+    Args:
+        text: The string to extract emails from
+        normalize: Whether to normalize emails by removing subaddresses
+                  and converting to lowercase
+
+    Returns:
+        A list of unique, validated email addresses
+    """
+    potential_emails = re.findall(EMAIL_REGEX, text)
+    unique_emails = set()
+
+    for email in potential_emails:
+        try:
+            # First validate the email with Pydantic
+            validated_email = EmailModel(email=email).email
+
+            # Apply normalization if requested
+            if normalize:
+                validated_email = normalize_email(validated_email)
+
+            unique_emails.add(validated_email)
+        except ValidationError:
+            pass
+
+    return list(unique_emails)
